Move acquisition functions to new files
