+++ Acquisition.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <OpenScanDeviceLib.h>
+
+#include <stdbool.h>
+
+OScDev_RichError *RunAcquisitionLoop(OScDev_Device *device);
+OScDev_RichError *StopAcquisitionAndWait(OScDev_Device *device);
+OScDev_RichError *IsAcquisitionRunning(OScDev_Device *device, bool *isRunning);
+OScDev_RichError *WaitForAcquisitionToFinish(OScDev_Device *device);
