+++ NIDAQ_OpenScanDevice.c
@@ -1,5 +1,6 @@
 #include "NIDAQ_OpenScanDevice.h"
 
+#include "Acquisition.h"
 #include "NIDAQ_OpenScanSettings.h"
 #include "OScNIDAQ.h"
 #include "OScNIDAQPrivateData.h"
